requirements.txt
absl-py==0.9.0
astor==0.8.1
bleach==1.5.0
gast==0.4.0
grpcio==1.31.0
h5py==2.10.0
html5lib==0.9999999
importlib-metadata==1.7.0
joblib==0.14.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
Markdown==3.2.2
nibabel==3.0.2
numpy==1.16.1
opencv-python==4.2.0.32
packaging==20.4
protobuf==3.13.0
pyparsing==2.4.7
PyYAML==5.3.1
scikit-learn==0.20.2
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
zipp==1.2.0
